:recycle: move router files
