--- src/store/drivers/index.js
@@ -1,5 +1,0 @@
-import router from './router'
-
-export default {
-  router,
-}
