--- src/store/drivers/router/index.js
@@ -1,8 +1,0 @@
-import { router } from '@k-redux-router/react-k-ramel'
-import routes from './routes'
-
-export default router({
-  routes,
-  state: 'ui.router',
-  getState: state => state.ui.router,
-})
