+++ src/store/store.js
@@ -1,8 +1,9 @@
 import { createStore } from 'k-ramel'
+import { router } from '@k-redux-router/react-k-ramel'
 
 import reducers from './reducers'
 import listeners from './listeners'
-import drivers from './drivers'
+import routes from './routes'
 
 /**
  * Store and dispatch actions are managed with k-ramel
@@ -10,7 +11,13 @@ import drivers from './drivers'
  */
 const store = createStore(reducers, {
   listeners,
-  drivers,
+  drivers: {
+    router: router({
+      routes,
+      state: 'ui.router',
+      getState: state => state.ui.router,
+    }),
+  },
   devtools: true,
 })
